PyDash/requirements.txt
asyncio==3.4.3
bleak==0.22.3
certifi==2024.8.30
charset-normalizer==3.4.0
dbus-fast==2.24.3
dbus-python==1.3.2
glib==1.0.0
idna==3.10
obd==0.7.2
packaging==24.1
Pint==0.20.1
plotly==5.24.1
pycairo==1.27.0
pygame==2.6.0
pygame-widgets==1.1.5
PyGObject==3.50.0
pyserial==3.5
requests==2.32.3
tenacity==9.0.0
urllib3==2.2.3
